AutonomousDiscordBot/requirements.txt
discord.py==2.1.0
aiosqlite==0.17.0
pyyaml==6.0
parsedatetime==2.6
pytest==7.3.1
pytest-asyncio==0.21.0
